+++ rv-predict-agent/src/main/java/com/runtimeverification/rvpredict/instrument/RVPredictInterceptor.java
@@ -1,4 +1,4 @@
-package com.runtimeverification.rvpredict.instrumentation;
+package com.runtimeverification.rvpredict.instrument;
 
 import java.util.ArrayList;
 import java.util.List;
