+++ rv-predict-jar/pom.xml
@@ -36,7 +36,7 @@
                             <mainClass>com.runtimeverification.rvpredict.engine.main.GUIMain</mainClass>
                         </manifest>
                         <manifestEntries>
-                            <Premain-Class>com.runtimeverification.rvpredict.instrumentation.Agent</Premain-Class>
+                            <Premain-Class>com.runtimeverification.rvpredict.instrument.Agent</Premain-Class>
                             <Can-Retransform-Classes>true</Can-Retransform-Classes>
                         </manifestEntries>
                     </archive>
